Add log serializer overloads (prevents a breaking change)
Changing the log serializer to support trimming means that we need to add the types that we will serialize into a serialization context. Unfortunately, the serializer is a public type that serializes arbitrary objects today. This is incompatible with trimming. The solution is to mark the existing API as [RequiresExternalCode], which gives trimmed callers of the API a warning, and to add overloads that accept the types that are in the  generation context.