Add dotnet core as a target framework
This avoids some dependency issues, since IdentityModel targets both netstandard and net6.0. In the IM netstandard build, we include the System.Text.Json dependency explicitly. In the IM net6.0 build, we don't, because that's built in. This would cause an issue if an application targeted net6.0 and depended on OidcClient. If OidcClient only targets netstandard, then when it depends on IM, we get the  netstandard build of IM, and bring along the System.Text.Json dependency.